requirements.txt
flask==3.0.3
flask_sqlalchemy==3.1.1
flask-wtf==1.2.1
email-validator==2.2.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
psycopg[binary]==3.2.3
pytest==8.3.3
pytest-flask==1.3.0